requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.3
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
